fix(sync): partial result should continue to reconcile
When there is a partial failure, lastErr was being set. This caused the
reconciler to stop reconciling conditions.

We should surface partial failures but not in a way that stops the
reconciler from applying the partial conditions.

When there is a partial result, omit the errors but still surface the
issue in the get_failures metric.